+++ CelestiumPT/src/shape_intersection.cuh
@@ -12,7 +12,7 @@ class Light;
 struct ShapeIntersection {
 	float hit_distance = -1;
 	int triangle_idx = -1;
-	long long int object_idx = -1;
+	int object_idx = -1;
 	float3 bary{};
 	float2 uv{};
 	float3 w_pos{};
@@ -26,6 +26,8 @@ struct ShapeIntersection {
 	Mat4 m_invModelMatrix;
 
 	bool front_face = true;
+
+	//debug data
 	float3 GAS_debug = make_float3(0);
 	int bbox_hit_count = 0;
 	int hit_count = 0;
